Made ints in for loops size_t

+++ space.hpp
@@ -19,7 +19,7 @@ private:
   space(const space &);
   space &operator=(const space &);
 
-  static std::map<std::string, std::map<std::string, std::string> > data;
+  static std::map<std::string, std::map<std::string, std::string>> data;
   static std::stringstream dstream, istream;
   static space s;
 
@@ -27,7 +27,7 @@ private:
   {
     std::ofstream file("data.json");
     file << "{\n";
-    for (std::map<std::string, std::map<std::string, std::string> >::iterator entry = data.begin(); entry != data.end(); entry++)
+    for (std::map<std::string, std::map<std::string, std::string>>::iterator entry = data.begin(); entry != data.end(); entry++)
     {
       file << "  "
            << "\"" << entry->first << "\""
@@ -1218,7 +1218,7 @@ public:
   static void sout(std::string name, std::string &type, int array[], int size)
   {
     std::string fixedType = "";
-    for (int i = 0; i < type.length(); ++i)
+    for (size_t i = 0; i < type.length(); ++i)
     {
       fixedType += char(std::tolower(type[i]));
     }
@@ -1231,7 +1231,7 @@ public:
       std::cout << "a 'stack', 'queue', or 'dequeue': ";
       std::cin >> fixedType;
 
-      for (int i = 0; i < type.length(); ++i)
+      for (size_t i = 0; i < type.length(); ++i)
       {
         fixedType += char(std::tolower(type[i]));
       }
@@ -1258,7 +1258,7 @@ public:
   static void sout(std::string name, std::string &type, double array[], int size)
   {
     std::string fixedType = "";
-    for (int i = 0; i < type.length(); ++i)
+    for (size_t i = 0; i < type.length(); ++i)
     {
       fixedType += char(std::tolower(type[i]));
     }
@@ -1271,7 +1271,7 @@ public:
       std::cout << "a 'stack', 'queue', or 'dequeue': ";
       std::cin >> fixedType;
 
-      for (int i = 0; i < type.length(); ++i)
+      for (size_t i = 0; i < type.length(); ++i)
       {
         fixedType += char(std::tolower(type[i]));
       }
@@ -1298,7 +1298,7 @@ public:
   static void sout(std::string name, std::string &type, char array[], int size)
   {
     std::string fixedType = "";
-    for (int i = 0; i < type.length(); ++i)
+    for (size_t i = 0; i < type.length(); ++i)
     {
       fixedType += char(std::tolower(type[i]));
     }
@@ -1311,7 +1311,7 @@ public:
       std::cout << "a 'stack', 'queue', or 'dequeue': ";
       std::cin >> fixedType;
 
-      for (int i = 0; i < type.length(); ++i)
+      for (size_t i = 0; i < type.length(); ++i)
       {
         fixedType += char(std::tolower(type[i]));
       }
@@ -1338,7 +1338,7 @@ public:
   static void sout(std::string name, std::string &type, bool array[], int size)
   {
     std::string fixedType = "";
-    for (int i = 0; i < type.length(); ++i)
+    for (size_t i = 0; i < type.length(); ++i)
     {
       fixedType += char(std::tolower(type[i]));
     }
@@ -1351,7 +1351,7 @@ public:
       std::cout << "a 'stack', 'queue', or 'dequeue': ";
       std::cin >> fixedType;
 
-      for (int i = 0; i < type.length(); ++i)
+      for (size_t i = 0; i < type.length(); ++i)
       {
         fixedType += char(std::tolower(type[i]));
       }
@@ -1378,7 +1378,7 @@ public:
   static void sout(std::string name, std::string &type, float array[], int size)
   {
     std::string fixedType = "";
-    for (int i = 0; i < type.length(); ++i)
+    for (size_t i = 0; i < type.length(); ++i)
     {
       fixedType += char(std::tolower(type[i]));
     }
@@ -1391,7 +1391,7 @@ public:
       std::cout << "a 'stack', 'queue', or 'dequeue': ";
       std::cin >> fixedType;
 
-      for (int i = 0; i < type.length(); ++i)
+      for (size_t i = 0; i < type.length(); ++i)
       {
         fixedType += char(std::tolower(type[i]));
       }
@@ -1418,7 +1418,7 @@ public:
   static void sout(std::string name, std::string &type, wchar_t array[], int size)
   {
     std::string fixedType = "";
-    for (int i = 0; i < type.length(); ++i)
+    for (size_t i = 0; i < type.length(); ++i)
     {
       fixedType += char(std::tolower(type[i]));
     }
@@ -1431,7 +1431,7 @@ public:
       std::cout << "a 'stack', 'queue', or 'dequeue': ";
       std::cin >> fixedType;
 
-      for (int i = 0; i < type.length(); ++i)
+      for (size_t i = 0; i < type.length(); ++i)
       {
         fixedType += char(std::tolower(type[i]));
       }
@@ -1905,7 +1905,7 @@ public:
     data[name]["data"] += "      " + dstream.str() + ",\n";
   }
 
-  static void sout(std::string name, std::vector<std::vector<int> > vector)
+  static void sout(std::string name, std::vector<std::vector<int>> vector)
   {
     data[name]["type"] = "2DVector";
     data[name]["is_char"] = "0";
@@ -1913,10 +1913,10 @@ public:
     istream << "[" << -1 << ", " << -1 << "]";
     dstream.str("");
     dstream << "[";
-    for (int i = 0; i < vector.size(); i++)
+    for (size_t i = 0; i < vector.size(); i++)
     {
       dstream << "[";
-      for (int j = 0; j < vector[i].size(); ++j)
+      for (size_t j = 0; j < vector[i].size(); ++j)
       {
         dstream << vector[i][j] << (j < vector[i].size() - 1 ? ", " : "");
       }
@@ -1928,7 +1928,7 @@ public:
     data[name]["data"] += "      " + dstream.str() + ",\n";
   }
 
-  static void sout(std::string name, std::vector<std::vector<float> > vector)
+  static void sout(std::string name, std::vector<std::vector<float>> vector)
   {
     data[name]["type"] = "2DVector";
     data[name]["is_char"] = "0";
@@ -1936,10 +1936,10 @@ public:
     istream << "[" << -1 << ", " << -1 << "]";
     dstream.str("");
     dstream << "[";
-    for (int i = 0; i < vector.size(); i++)
+    for (size_t i = 0; i < vector.size(); i++)
     {
       dstream << "[";
-      for (int j = 0; j < vector[i].size(); ++j)
+      for (size_t j = 0; j < vector[i].size(); ++j)
       {
         dstream << vector[i][j] << (j < vector[i].size() - 1 ? ", " : "");
       }
@@ -1951,7 +1951,7 @@ public:
     data[name]["data"] += "      " + dstream.str() + ",\n";
   }
 
-  static void sout(std::string name, std::vector<std::vector<double> > vector)
+  static void sout(std::string name, std::vector<std::vector<double>> vector)
   {
     data[name]["type"] = "2DVector";
     data[name]["is_char"] = "0";
@@ -1959,10 +1959,10 @@ public:
     istream << "[" << -1 << ", " << -1 << "]";
     dstream.str("");
     dstream << "[";
-    for (int i = 0; i < vector.size(); i++)
+    for (size_t i = 0; i < vector.size(); i++)
     {
       dstream << "[";
-      for (int j = 0; j < vector[i].size(); ++j)
+      for (size_t j = 0; j < vector[i].size(); ++j)
       {
         dstream << vector[i][j] << (j < vector[i].size() - 1 ? ", " : "");
       }
@@ -1974,7 +1974,7 @@ public:
     data[name]["data"] += "      " + dstream.str() + ",\n";
   }
 
-  static void sout(std::string name, std::vector<std::vector<bool> > vector)
+  static void sout(std::string name, std::vector<std::vector<bool>> vector)
   {
     data[name]["type"] = "2DVector";
     data[name]["is_char"] = "0";
@@ -1982,10 +1982,10 @@ public:
     istream << "[" << -1 << ", " << -1 << "]";
     dstream.str("");
     dstream << "[";
-    for (int i = 0; i < vector.size(); i++)
+    for (size_t i = 0; i < vector.size(); i++)
     {
       dstream << "[";
-      for (int j = 0; j < vector[i].size(); ++j)
+      for (size_t j = 0; j < vector[i].size(); ++j)
       {
         dstream << vector[i][j] << (j < vector[i].size() - 1 ? ", " : "");
       }
@@ -1997,7 +1997,7 @@ public:
     data[name]["data"] += "      " + dstream.str() + ",\n";
   }
 
-  static void sout(std::string name, std::vector<std::vector<char> > vector)
+  static void sout(std::string name, std::vector<std::vector<char>> vector)
   {
     data[name]["type"] = "2DVector";
     data[name]["is_char"] = "1";
@@ -2005,10 +2005,10 @@ public:
     istream << "[" << -1 << ", " << -1 << "]";
     dstream.str("");
     dstream << "[";
-    for (int i = 0; i < vector.size(); i++)
+    for (size_t i = 0; i < vector.size(); i++)
     {
       dstream << "[";
-      for (int j = 0; j < vector[i].size(); ++j)
+      for (size_t j = 0; j < vector[i].size(); ++j)
       {
         dstream << (int)vector[i][j] << (j < vector[i].size() - 1 ? ", " : "");
       }
@@ -2020,7 +2020,7 @@ public:
     data[name]["data"] += "      " + dstream.str() + ",\n";
   }
 
-  static void sout(std::string name, std::vector<std::vector<wchar_t> > vector)
+  static void sout(std::string name, std::vector<std::vector<wchar_t>> vector)
   {
     data[name]["type"] = "2DVector";
     data[name]["is_char"] = "0";
@@ -2028,10 +2028,10 @@ public:
     istream << "[" << -1 << ", " << -1 << "]";
     dstream.str("");
     dstream << "[";
-    for (int i = 0; i < vector.size(); i++)
+    for (size_t i = 0; i < vector.size(); i++)
     {
       dstream << "[";
-      for (int j = 0; j < vector[i].size(); ++j)
+      for (size_t j = 0; j < vector[i].size(); ++j)
       {
         dstream << vector[i][j] << (j < vector[i].size() - 1 ? ", " : "");
       }
@@ -2044,7 +2044,7 @@ public:
   }
 };
 
-std::map<std::string, std::map<std::string, std::string> > space::data;
+std::map<std::string, std::map<std::string, std::string>> space::data;
 std::stringstream space::dstream, space::istream;
 space space::s;
